Updated styling for the work order forms.
Completed a drop down listing for the status and category on the create form and workorder controller. This drop down list currently shows the ids of those objects since the workorder needs those ids to be created. Still working on the employee finding method using the user repostitory and service. I will need to modify the outputs on the views so that names can be read by the users, but the controller will recieve the ids of those objects in order to properly create or edit the work order object.